修改U1,U2...U8的Java类型代表，并为其添加注释

+++ src/test/java/buddha/jvmabi/ClassFileBasicInfoTest.java
@@ -27,8 +27,8 @@ public class ClassFileBasicInfoTest extends BaseTest {
         assertEquals(ClassFileConstantTagConst.NAME_AND_TYPE_INFO, pool.get(3).getTag());
         assertEquals(ClassFileConstantTagConst.UTF8_INFO, pool.get(6).getTag());
         assertEquals(ClassFileConstantTagConst.UTF8_INFO, pool.get(24).getTag());
-        final short nameIndex = pool.<ConstantNameAndType>getExact(9).getNameIndex();
-        final short descIndex = pool.<ConstantNameAndType>getExact(9).getDescriptorIndex();
+        final int nameIndex = pool.<ConstantNameAndType>getExact(9).getNameIndex();
+        final int descIndex = pool.<ConstantNameAndType>getExact(9).getDescriptorIndex();
         assertEquals(11, nameIndex);
         assertEquals(12, descIndex);
         for (int i = 1, len = pool.length(); i < len; i++) {
